move urlreader to domain, add test for createFromUrl
