+++ app/src/main/java/lukuvinkkikirjasto/domain/UrlReaderWrapper.java
@@ -1,4 +1,4 @@
-package lukuvinkkikirjasto.dao; 
+package lukuvinkkikirjasto.domain; 
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import java.io.IOException;
