+++ app/src/test/java/lukuvinkkikirjasto/domain/FakeUrlReader.java
@@ -1,0 +1,8 @@
+package lukuvinkkikirjasto.domain;
+import java.io.IOException;
+
+class FakeUrlReader implements UrlReaderProvider {
+    public String getTitleFrom(String url) throws IOException {
+        return "FakeTitle";
+    } 
+}
